backend/requirements.txt
Django==3.2.13
django-filter==22.1
djangorestframework==3.13.1
drf-base64==2.0
djoser==2.1.0
psycopg2-binary==2.9.3
backports.zoneinfo==0.2.1
importlib-metadata==1.7.0
Markdown==3.3.4
Pillow==9.2.0
python-dotenv==0.20.0
gunicorn==20.1.0
flake8==5.0.4
